Added table countdown + some fixes
